Enklere tilnærming, bruk metode med checked exception når nøkkel mangler
